git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@11039 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/org/starexec/data/to/JobPair.java
@@ -23,7 +23,7 @@ public class JobPair extends Identifiable {
 	private Timestamp startTime = null;
 	private Timestamp endTime = null;	
 	private int exitStatus;
-	private long wallclockTime;
+	private double wallclockTime;
 	private double cpuUsage;
 	private double userTime;
 	private double systemTime;
@@ -242,7 +242,7 @@ public class JobPair extends Identifiable {
 	/**
 	 * @return the wallclock time it took for this pair to execute in seconds
 	 */
-	public long getWallclockTime() {
+	public double getWallclockTime() {
 		return wallclockTime;
 	}
